Fixed some slight bugs

+++ example/DemoSources/Landscape.hpp
@@ -32,7 +32,7 @@ namespace Landscape
 
 		const std::int_fast32_t XPos{ 63 + (((static_cast<std::int_fast32_t>(std::cosf(LandscapeAngle * 0.01F) * 128.0F)) << 7) >> 7) };
 		const std::int_fast32_t ZPos{ 63 + (((static_cast<std::int_fast32_t>(std::sinf(LandscapeAngle * 0.01F) * 128.0F)) << 7) >> 7) };
-		const float Factor{ ZOrigin - ZPos > 0 ? 100.0F * (std::atanf(static_cast<float>((XOrigin - XPos)) / (ZOrigin - ZPos))) : 100.0F * (std::acosf(-1.0F) + std::atanf(static_cast<float>((XOrigin - XPos)) / (ZOrigin - ZPos))) };
+		const float Factor{ static_cast<float>(ZOrigin - ZPos) > 0.0F ? 100.0F * (std::atanf(static_cast<float>((XOrigin - XPos)) / static_cast<float>(ZOrigin - ZPos))) : 100.0F * (std::acosf(-1.0F) + std::atanf(static_cast<float>((XOrigin - XPos)) / static_cast<float>(ZOrigin - ZPos))) };
 		const std::int_fast32_t CosA{ static_cast<std::int_fast32_t>(std::cosf(Factor * 0.01F) * 128.0F) };
 		const std::int_fast32_t SinA{ static_cast<std::int_fast32_t>(std::sinf(Factor * 0.01F) * 128.0F) };
